+++ scripts/macros/2024/spells/callLightning.js
@@ -1,7 +1,5 @@
 import {callLightning as callLightningLegacy} from '../../2014/spells/callLightning.js';
 export let callLightning = {
-    name: 'Call Lightning',
-    version: '1.2.28',
-    rules: 'modern',
-    midi: callLightningLegacy.midi
+    ...callLightningLegacy,
+    rules: 'modern'
 };